+++ src/api/controllers/search/index.spec.js
@@ -1,4 +1,4 @@
-import * as testData from 'test/testData';
+import * as testData from 'test/testData/db';
 import controller from './';
 
 describe('search controller', () => {