+++ CHANGELOG.md
@@ -8,6 +8,13 @@
 - Changes
 ```
 
+## [2.0.0] - 25/02/17
+### Added
+- Wins/losses to pvp standings
+
+### Removed
+- Properties from leaderboard endpoints that weren't needed
+
 ## [1.7.1] - 20/02/17
 ### Fixed
 - Local environment from not being able to start up correctly